Provide a faster and more accurate inverse matrix
Faster variant is implemented for affine matrices. This resolves also
an issue that for general inversion of matrices, the 1 in the lower
right corner tends to drift as iterations continue.

For perspective matrices standard inv() is called. To resolve the drift
mentioned above, one could rewrite the inverse compositional as

            Warp<WARP_SIMILARITY> wDelta;
            wDelta.setParameters(delta);
            wDelta.setMatrix(wDelta.invMatrix());
            wDelta.setParameters(wDelta.parameters()); // this should
enforce a 1 in the lower right.
            setMatrix(matrix() * wDelta.matrix());